+++ java/java-impl/src/com/intellij/codeInsight/daemon/impl/quickfix/AccessStaticViaInstanceFix.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2009 JetBrains s.r.o.
+ * Copyright 2000-2014 JetBrains s.r.o.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -97,9 +97,11 @@ public class AccessStaticViaInstanceFix extends LocalQuickFixAndIntentionActionO
         if (!checkSideEffects(project, containingClass, qualifierExpression, factory, myExpression,editor)) return;
         PsiElement newQualifier = qualifierExpression.replace(factory.createReferenceExpression(containingClass));
         PsiElement qualifiedWithClassName = myExpression.copy();
-        newQualifier.delete();
-        if (myExpression.resolve() != myMember) {
-          myExpression.replace(qualifiedWithClassName);
+        if (myExpression.getTypeParameters().length == 0) {
+          newQualifier.delete();
+          if (myExpression.resolve() != myMember) {
+            myExpression.replace(qualifiedWithClassName);
+          }
         }
       }
     }
@@ -163,9 +165,9 @@ public class AccessStaticViaInstanceFix extends LocalQuickFixAndIntentionActionO
       };
     dialog.show();
     int res = dialog.getExitCode();
-    if (res == RemoveUnusedVariableUtil.CANCEL) return false;
+    if (res == RemoveUnusedVariableUtil.RemoveMode.CANCEL.ordinal()) return false;
     try {
-      if (res == RemoveUnusedVariableUtil.MAKE_STATEMENT) {
+      if (res == RemoveUnusedVariableUtil.RemoveMode.MAKE_STATEMENT.ordinal()) {
         final PsiStatement statementFromText = factory.createStatementFromText(qualifierExpression.getText() + ";", null);
         final PsiStatement statement = PsiTreeUtil.getParentOfType(myExpression, PsiStatement.class);
         statement.getParent().addBefore(statementFromText, statement);
