+++ java/java-impl/src/com/intellij/refactoring/safeDelete/usageInfo/SafeDeleteReferenceJavaDeleteUsageInfo.java
@@ -16,8 +16,9 @@
 package com.intellij.refactoring.safeDelete.usageInfo;
 
 import com.intellij.openapi.diagnostic.Logger;
-import com.intellij.psi.PsiElement;
+import com.intellij.psi.*;
 import com.intellij.refactoring.safeDelete.ImportSearcher;
+import com.intellij.refactoring.util.RefactoringUtil;
 import com.intellij.util.IncorrectOperationException;
 
 /**
@@ -45,10 +46,21 @@ public class SafeDeleteReferenceJavaDeleteUsageInfo extends SafeDeleteReferenceS
       LOG.assertTrue(element != null);
       PsiElement importStatement = ImportSearcher.getImport(element, false);
       if (importStatement != null) {
-        importStatement.delete();
+        if (element instanceof PsiImportStaticReferenceElement) {
+          if (((PsiImportStaticReferenceElement)element).multiResolve(false).length < 2) {
+            importStatement.delete();
+          }
+        } else {
+          importStatement.delete();
+        }
       }
       else {
-        element.delete();
+        if (element instanceof PsiExpressionStatement && RefactoringUtil.isLoopOrIf(element.getParent())) {
+          final PsiStatement emptyTest = JavaPsiFacade.getInstance(getProject()).getElementFactory().createStatementFromText(";", null);
+          element.replace(emptyTest);
+        } else {
+          element.delete();
+        }
       }
     }
   }
